systemd: Combine the server.js journalctl.js and renderer.js
Refine the internal API here, and cleanup unneeded exported
functions.
