+++ pkg/storaged/overview.js
@@ -22,12 +22,12 @@ define([
     "base1/cockpit",
     "./mustache",
     "./plot",
-    "system/server",
+    "system/journal",
     "storage/utils",
     "storage/dialog",
     "storage/permissions",
     "./flot",
-], function($, cockpit, mustache, plot, server, utils, dialog, permissions) {
+], function($, cockpit, mustache, plot, journal, utils, dialog, permissions) {
     var _ = cockpit.gettext;
     var C_ = cockpit.gettext;
 
@@ -399,7 +399,7 @@ define([
         render_jobs();
 
         $('#storage-log').append(
-            server.logbox([ "_SYSTEMD_UNIT=storaged.service", "+",
+            journal.logbox([ "_SYSTEMD_UNIT=storaged.service", "+",
                             "_SYSTEMD_UNIT=udisks2.service", "+",
                             "_SYSTEMD_UNIT=dm-event.service", "+",
                             "_SYSTEMD_UNIT=smartd.service", "+",
