+++ pkg/systemd/logs.html
@@ -24,7 +24,7 @@ along with Cockpit; If not, see <http://www.gnu.org/licenses/>.
   <meta charset="utf-8">
   <link href="../base1/cockpit.css" rel="stylesheet">
   <link href="host.css" rel="stylesheet">
-  <script type="text/javascript" data-main="system/journal" src="../base1/bundle.js"></script>
+  <script type="text/javascript" data-main="system/logs" src="../base1/bundle.js"></script>
   <script type="text/javascript" src="bundle.js"></script>
 </head>
