+++ po/POTFILES.in
@@ -21,9 +21,8 @@ pkg/systemd/patterns.js
 pkg/systemd/plot.js
 pkg/systemd/journal.js
 pkg/systemd/init.js
-pkg/systemd/journalctl.js
-pkg/systemd/server.js
-pkg/systemd/renderer.js
+pkg/systemd/logs.js
+pkg/systemd/journal.js
 pkg/systemd/host.js
 pkg/realmd/operation.js
 pkg/docker/overview.js
